Add a control VM - WSL2 may be causing problems
I've been using WSL2 as the Ansible control host and I think it may be
causing performance problems that are severe enough to break the
cluster. WSL2 is backed by a level 1 Hyper-V vm, while VirtualBox is
using level 2 KVM VMs. I think this is causing contention for VT
extensions. The main issue is that flannel Pods keep crashing because
the nodes lock-up frequently. Here's hoping that using a fourth VM
(control) for the Ansible control host solves the problem.