Download kube-flannel.yml, update before install
Thanks to StackOverflow and ChatGPT, I found that my VirtualBox VMs, as
created by the Vagrantfile that adds sets up a private network, and
causing the VMs to sprout a second nic, I needed to specify that nice,
"eth1" in this case, when installing kube-flannel.

Thus, this update downloads kube-flannel.yml and adds '- --iface="eth1"'
to the container args.